Common EditServerView logic
